+++ app/src/main/java/bootcamp/sparta/nb_deepen_assignment/ui/main/MainViewPagerAdapter.kt
@@ -1,9 +1,11 @@
-package bootcamp.sparta.nb_deepen_assignment.ui
+package bootcamp.sparta.nb_deepen_assignment.ui.main
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import bootcamp.sparta.nb_deepen_assignment.R
+import bootcamp.sparta.nb_deepen_assignment.ui.mypage.MyPageFragment
+import bootcamp.sparta.nb_deepen_assignment.ui.search.SearchFragment
 
 class MainViewPagerAdapter(fragmentActivity: FragmentActivity): FragmentStateAdapter(fragmentActivity){
     private val list = ArrayList<MainTab>()
@@ -24,6 +26,8 @@ class MainViewPagerAdapter(fragmentActivity: FragmentActivity): FragmentStateAda
     }
 
     fun getTabName(position: Int): Int = list[position].tabNameRes
+    fun getSearchFragment() : SearchFragment? = list.find { it.fragment is SearchFragment }?.fragment as? SearchFragment
+    fun getMyPageFragment() : MyPageFragment? = list.find { it.fragment is MyPageFragment }?.fragment as? MyPageFragment
 
     override fun getItemCount(): Int {
         return list.size
