+++ app/src/main/res/layout/search_item.xml
@@ -8,7 +8,7 @@
     xmlns:tool="http://schemas.android.com/tools">
 
    <ImageView
-       android:id="@+id/image_mypage"
+       android:id="@+id/image"
        android:layout_width="wrap_content"
        android:layout_height="100dp"
        tool:src="@drawable/ic_launcher_background"
@@ -16,24 +16,24 @@
        app:layout_constraintTop_toTopOf="parent" />
 
    <TextView
-      android:id="@+id/dttm_mypage"
+      android:id="@+id/dttm"
        android:layout_width="0dp"
        android:layout_height="wrap_content"
-      app:layout_constraintStart_toEndOf="@+id/image_mypage"
-       app:layout_constraintEnd_toStartOf="@+id/like_mypage"
+      app:layout_constraintStart_toEndOf="@+id/image"
+       app:layout_constraintEnd_toStartOf="@+id/like"
       app:layout_constraintTop_toTopOf="parent"
        app:layout_constraintBottom_toBottomOf="parent"
        />
 
    <androidx.appcompat.widget.AppCompatButton
-      android:id="@+id/like_mypage"
+      android:id="@+id/like"
        android:layout_width="40dp"
        android:layout_height="40dp"
        android:background="@drawable/baseline_star_border_24"
-       app:layout_constraintStart_toEndOf="@+id/dttm_mypage"
+       app:layout_constraintStart_toEndOf="@+id/dttm"
        app:layout_constraintEnd_toEndOf="parent"
-       app:layout_constraintTop_toTopOf="@+id/dttm_mypage"
-       app:layout_constraintBottom_toBottomOf="@+id/dttm_mypage"
+       app:layout_constraintTop_toTopOf="@+id/dttm"
+       app:layout_constraintBottom_toBottomOf="@+id/dttm"
        />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
